+++ caddy2/rootfs/etc/s6-overlay/s6-rc.d/init/00-banner.sh
@@ -1,0 +1,7 @@
+#!/command/with-contenv sh
+
+echo "==========================================="
+cat /etc/s6-overlay/s6-rc.d/init/00-banner
+echo " "
+echo "Caddy版本：CADDY_VER"
+echo "==========================================="